+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 - Nothing yet
+
+## [3.3.1] - 2024-09-23
+### Fixed
+- Fixed a bug in CsvReader where lines were mistakenly treated as empty and skipped when skipEmptyLines was set (default). These affected lines made up solely of field separators, solely empty quoted fields, or fields rendered empty after applying optional field modifiers.
 
 ## [3.3.0] - 2024-09-19
 ### Added
@@ -126,7 +131,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/osiegmar/FastCSV/compare/v3.3.0...main
+[Unreleased]: https://github.com/osiegmar/FastCSV/compare/v3.3.1...main
+[3.3.0]: https://github.com/osiegmar/FastCSV/compare/v3.3.0...v3.3.1
 [3.3.0]: https://github.com/osiegmar/FastCSV/compare/v3.2.0...v3.3.0
 [3.2.0]: https://github.com/osiegmar/FastCSV/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/osiegmar/FastCSV/compare/v3.0.0...v3.1.0